requirements.txt
# Existing
requests>=2.31
fastapi>=0.110
uvicorn>=0.29
pandas>=2.0
pdfplumber>=0.10
openpyxl>=3.1

# New — Data processing
spacy>=3.7
rapidfuzz>=3.5
recordlinkage>=0.15
python-dateutil>=2.9

# New — Database
sqlalchemy>=2.0
psycopg2-binary>=2.9

# New — Flask API
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
bcrypt>=4.1
passlib>=1.7

# New — ML
scikit-learn>=1.3
statsmodels>=0.14
joblib>=1.3

# New — Scheduling
apscheduler>=3.10

# New — Utils
python-dotenv>=1.0
pytest>=7.0
httpx>=0.27
